refactor(commit): bin配下のshebangスクリプト経由実行に切り替え
スキルの`allowed-tools`に指定していた`Bash(node:*)`は許可するコマンドが広すぎたため、
prプラグインと同じ方針に揃えて、
`bin/`配下にshebang付きで実行可能配置したスクリプトを直接呼び出す形に変更します。

`bin/commit-prepare.ts`と`bin/read-commit-instructions.ts`に`#!/usr/bin/env node`を付与して実行権限を立て、
`package.json`の`bin`セクションにも登録することで、
スキルからは`!`commit-prepare.ts``と`!`read-commit-instructions.ts``として呼び出せるようにします。
これに合わせて`allowed-tools`からは`Bash(node:*)`を外し、
`Bash(commit-prepare.ts:*)`と`Bash(read-commit-instructions.ts:*)`に絞り込みます。

ロジック部分は`bin/`から`src/commit-prepare.ts`と`src/read-commit-instructions.ts`に分離し、
`bin/`配下はshebang付きのthin wrapperだけにしました。
これに伴い旧`lib/commit-prepare.ts`と`lib/`ディレクトリは撤去し、
テストの`import`先も`../lib/`から`../src/`に書き換えます。

プラグインバージョンを`2.6.2`から`2.6.3`に上げます。
