+++ plugins/commit/.claude-plugin/plugin.json
@@ -1,6 +1,6 @@
 {
   "name": "commit",
-  "version": "2.6.2",
+  "version": "2.6.3",
   "description": "Generate a commit message from staged changes and let the user review before committing. Use when the user wants to commit changes or create a git commit.",
   "author": {
     "name": "ncaq",
